pyproject.toml
[project]
name = "pyopt"
version = "0.2.0"
description = "Stock Portfolio Optimization using Modern Portfolio Theory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair",
    "fastapi",
    "fastmcp",
    "matplotlib",
    "numpy",
    "pandas",
    "pyportfolioopt",
    "riskfolio-lib",
    "streamlit[starlette]",
    "vnstock",
]

[project.scripts]
pyopt = "pyopt_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["pyopt_cli.py", "streamlit_app.py", "app.py", "server.py", "backend", "frontend"]

[dependency-groups]
dev = ["ruff"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
